requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
alpaca-eval==0.3.6
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.1.0
blessed==1.20.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
deepspeed>=0.12.6
defusedxml==0.7.1
diffusers==0.24.0
dill==0.3.7
distro==1.8.0
docker-pycreds==0.4.0
et-xmlfile==1.1.0
fastapi==0.104.1
fastjsonschema==2.19.0
ffmpy==0.3.1
filelock==3.13.1
fire==0.5.0
flash-attn
fonttools==4.45.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gpustat==1.1.1
gradio==3.35.0
gradio_client==0.7.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.20.3
icecream==2.1.3
idna==3.4
importlib-metadata==7.0.0
importlib-resources==6.1.1
iopath==0.1.10
ipywidgets==8.1.1
isoduration==20.11.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
Levenshtein==0.24.0
linkify-it-py==2.0.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
networkx==3.0
ninja==1.11.1.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
nvidia-ml-py==12.535.133
nvitop==1.3.1
omegaconf==2.3.0
openai==1.3.5
opencv-python==4.9.0.80
openpyxl==3.1.2
orjson==3.9.10
overrides==7.4.0
pandas==2.1.3
pandas-compat==0.1.1
pandocfilters==1.5.0
Pillow==9.3.0
plotly==5.19.0
portalocker==2.8.2
prometheus-client==0.19.0
protobuf==4.25.1
psutil
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pydub==0.25.1
pynvml==11.5.0
pyparsing==3.1.1
pypinyin==0.50.0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-Levenshtein==0.24.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.6.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rouge==1.0.1
rpds-py==0.13.1
runipy==0.1.5
safetensors==0.4.2
scipy==1.11.4
semantic-version==2.10.0
Send2Trash==1.8.2
sentry-sdk==1.35.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.11
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tiktoken==0.5.1
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
tornado==6.4
transformers==4.37.0
transformers-stream-generator==0.0.4
triton
typer==0.9.0
types-python-dateutil==2.8.19.14
tzdata==2023.3
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.24.0.post1
wandb==0.16.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
widgetsnbextension==4.0.9
word2number==1.1
xformers==0.0.29.post3
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
